Factor out subtree similarity computations.
