+++ src/lib/matchers.rs
@@ -38,7 +38,7 @@
 #![warn(missing_docs)]
 
 use std::cell::RefCell;
-use std::collections::{HashMap, HashSet};
+use std::collections::HashMap;
 use std::fmt::Debug;
 
 use action::EditScript;
@@ -223,66 +223,6 @@ impl<T: Clone + Debug + Eq + ToString + 'static> MappingStore<T> {
         self.from_arena.borrow()[*from].ty == self.to_arena.borrow()[*to].ty
         && !(self.contains_from(from) || self.contains_to(to))
     }
-
-    /// Dice measure of similarity between subtrees.
-    pub fn dice_sim(&self, from: &NodeId<FromNodeId>, to: &NodeId<ToNodeId>) -> f64 {
-        let n_from = from.breadth_first_traversal(&self.from_arena.borrow())
-                         .collect::<Vec<NodeId<FromNodeId>>>()
-                         .len() as f64;
-        let n_to =
-            to.breadth_first_traversal(&self.to_arena.borrow()).collect::<Vec<NodeId<ToNodeId>>>()
-              .len() as f64;
-        let dice = 2.0 * f64::from(self.num_common_descendants(from, to)) / (n_from + n_to);
-        debug_assert!(dice >= 0. && dice <= 1.);
-        dice
-    }
-
-    /// Jaccard measure of similarity between subtrees.
-    pub fn jaccard_sim(&self, from: &NodeId<FromNodeId>, to: &NodeId<ToNodeId>) -> f64 {
-        let n_from = from.breadth_first_traversal(&self.from_arena.borrow())
-                         .collect::<Vec<NodeId<FromNodeId>>>()
-                         .len() as f64;
-        let n_to =
-            to.breadth_first_traversal(&self.to_arena.borrow()).collect::<Vec<NodeId<ToNodeId>>>()
-              .len() as f64;
-        let common = f64::from(self.num_common_descendants(from, to));
-        let jaccard = common / (n_from + n_to - common);
-        debug_assert!(jaccard >= 0. && jaccard <= 1.);
-        jaccard
-    }
-
-    /// Measure of similarity between subtrees Described in Chawathe et al. (1996).
-    pub fn chawathe_sim(&self, from: &NodeId<FromNodeId>, to: &NodeId<ToNodeId>) -> f64 {
-        let n_from = from.breadth_first_traversal(&self.from_arena.borrow())
-                         .collect::<Vec<NodeId<FromNodeId>>>()
-                         .len() as f64;
-        let n_to =
-            to.breadth_first_traversal(&self.to_arena.borrow()).collect::<Vec<NodeId<ToNodeId>>>()
-              .len() as f64;
-        let common = f64::from(self.num_common_descendants(from, to));
-        let chawathe = common / n_from.max(n_to);
-        debug_assert!(chawathe >= 0. && chawathe <= 1.);
-        chawathe
-    }
-
-    /// Find the number of "common" descendants in two matched subtrees.
-    ///
-    /// To nodes are common if they have already been matched.
-    fn num_common_descendants(&self, from: &NodeId<FromNodeId>, to: &NodeId<ToNodeId>) -> u32 {
-        let mut dst_desc = HashSet::new();
-        for node in to.breadth_first_traversal(&self.to_arena.borrow()) {
-            dst_desc.insert(node);
-        }
-        let mut common = 0;
-        let mut to: Option<NodeId<ToNodeId>>;
-        for node in from.descendants(&self.from_arena.borrow()) {
-            to = self.get_to(&node);
-            if to.is_some() && dst_desc.contains(&to.unwrap()) {
-                common += 1;
-            }
-        }
-        common
-    }
 }
 
 /// Match two trees and return a store of mappings between them.
@@ -455,22 +395,6 @@ mod tests {
         assert!(!store.is_mapped(&NodeId::new(2), &NodeId::new(1)));
         assert!(!store.is_mapped(&NodeId::new(2), &NodeId::new(2)));
         assert!(!store.is_mapped(&NodeId::new(2), &NodeId::new(3)));
-    }
-
-    #[test]
-    fn num_common_descendants() {
-        let mult = create_mult_arena();
-        let plus = create_plus_arena();
-        let store = MappingStore::new(plus, Arena::<&'static str, ToNodeId>::from(mult));
-        store.push(NodeId::new(0), NodeId::new(2), &Default::default());
-        store.push(NodeId::new(1), NodeId::new(3), &Default::default());
-        store.push(NodeId::new(2), NodeId::new(4), &Default::default());
-        assert_eq!(2,
-                   store.num_common_descendants(&NodeId::new(0), &NodeId::new(2)));
-        assert_eq!(2,
-                   store.num_common_descendants(&NodeId::new(0), &NodeId::new(0)));
-        assert_eq!(0,
-                   store.num_common_descendants(&NodeId::new(1), &NodeId::new(0)));
     }
 
     #[test]
